AI_Worker/requirements.txt
numpy==1.24.3
onnxruntime==1.15.1
Pillow==10.0.0
psycopg2-binary==2.9.7
pika==1.3.2